feat: add nft data buffer and move is_printable helper
